01-Build-a-simple-AI-Agent/pyproject.toml
[project]
name = "01-build-a-simple-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6.4",
    "python-dotenv>=1.2.1",
]
